requirements.txt
# Requires Python >= 3.13
mcp[cli]
boto3
python-dotenv
